requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
attrs==24.2.0
autopep8==2.3.1
backoff==2.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
dill==0.3.8
distro==1.9.0
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.114
litellm==1.44.17
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mem0ai==0.1.48
monotonic==1.6
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.24.0
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.6.3
protobuf==5.28.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.29.0
pylint==3.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qdrant-client==1.11.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
setuptools==74.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomlkit==0.13.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.11
zipp==3.20.1
google-generativeai==0.8.4